Solves part2 by using LCM

+++ src/08-haunted-wasteland.h
@@ -8,5 +8,5 @@ class Solution_08
 {
     public:
     int part1(string instructions, map<string,pair<string,string>> network);
-    int part2(string instructions, map<string,pair<string,string>> network);
+    long part2(string instructions, map<string,pair<string,string>> network);
 };
